change some code

+++ src/main/java/com/example/demo/Controler/RatingTitleController.java
@@ -26,7 +26,7 @@ public class RatingTitleController {
     }
 
     @GetMapping("getAverageRating")
-    public RatingTitle getAverageRating(@RequestParam Integer doubleNum){
+    public RatingTitle getAverageRating(@RequestParam Double doubleNum){
         return (RatingTitle) ratingRepo.findByAverageRating(doubleNum);
     }
